Add additional display data, Conditionally render authDetails
I needed to have the auth details display the authinfo particular to the search. I was mapping over all dataNodes and pickeing a specific one, I changed auth details to
conditionally render the specific (target) node. Currently the details are displaying, but they are happening onchange, now I need it to display onclick.