Add strong tags to auth details
Included some strong tags to show difference between the title of each row vs the details that populate. Added an edit button and icon for editing info and some styles for them.
Also removed some unused styles and changed a span to a div in the patient info to keep semantics the same.