Remove hardcoded strings from action and reducer
Type was hardcoded stringin the authadmin action. Created an actiontypes to store action type variables. Imported action types into both the reducer and the action to have access to actiontype variables instead of hardcoded strings. Also moved mapdispatchToProps property to the name actions so that it doesnt need to be changed if more actions are added.