Make authdetails uneditable until edit button clicked
There needed to be an action before auth details can be edited, now it is not clickable until the edit button is clicked. Had some props failing in auth details. a couple
of the prop name changes were missed from when auth details was changed to a class. Did some format clean up along the way in action types and company component as well as
removed a un-needed prop in company component.