Add typeahead to company
Boot strap react typeahead npm package was brought in a while back, but a decision was made not to use it and to implement custom functionality. Turns out that the typeahead
bootstrap module is more than likely  a better option. This commit is the first implementation and is inside of the company component. More to follow.