Add thunk middleware
Added thunk middleware in and created one in reason for visit acrtions. Added seceral action types into action types and moved test reason data intp public folder to maock
the fetch call. Needed to update the reason for visit select value when changed and needed to populate it based off of an authorization search as well. It now loads when the
component mounts so it is ready for a search or change reason action.