Clean up and perform renaming. Begin working on other components in visit info starting with date of visit
Reason for visit was complete, performed clean up before moving on to the other components in visit info.
Check in status's action and reducer had some naming conflicts. Changed statuses to status to match the rest of the code.
Removed folders that had components which are not being used and are not going to be used.
Began work on date of visit component. Date picket date will need to be reformated to match the date that sits in the db.
Date reformat will need to happen in an action, created auth details action.
Added a save button to fire authDetails action, and functionality to make edit and save buttons hide and show when the other is clicked.