Set up scaffolding for auth details action thunk
Date picker date is now passing through moment and coming back very close to the format it will need to be in to be put into the databaase.
Mapped dispatch to props in auth details and it is firing properly.
Gave the set date of visit action a date variable from date picker instead of pulling from props. This was the date that clean auth details data needed to correctly format inside of moment